changed the key_size and elt_size parameters to size_t in ht-mul-uint64

+++ data-structures/ht-mul-uint64/ht-mul-uint64.h
@@ -81,8 +81,8 @@ typedef struct{
                  element as its parameter, is necessary to delete the element
 */
 void ht_mul_uint64_init(ht_mul_uint64_t *ht,
-                        int key_size,
-	                int elt_size,
+                        size_t key_size,
+	                size_t elt_size,
 			float alpha,
                         int (*cmp_key)(const void *, const void *),
                         void (*rdc_key)(void *, const void *),
